refactor: 중복되는 로직 커스텀 훅으로 분리
