+++ src/containers/BookmarkContainer.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import {
@@ -10,7 +10,7 @@ import {
 	removeAllBookmarkItems,
 } from "../slice";
 import popup from "../popup";
-import { usePreviewItem } from "../hooks";
+import { usePreviewItem, useSearch } from "../hooks";
 import { CUSTOM_COMMAND_TAB } from "../constants/sidebarTab";
 import CharacterContainerLayout from "../layouts/CharacterContainerLayout";
 import SideTabItemLayout from "../layouts/SideTabItemLayout";
@@ -24,7 +24,7 @@ import CharacterListItem from "../presentationals/CharacterListItem";
 export default function BookmarkContainer({ setTabState }) {
 	const dispatch = useDispatch();
 	const { bookmarkItems, latexInput } = useSelector(state => state);
-	const [searchTerm, setSearchTerm] = useState("");
+	const [searchTerm, handleFilter] = useSearch("");
 	const [previewItem, handleMouseEnterItem] = usePreviewItem({ id: "", top: 0 });
 
 	const handleCustomButtonClick = latex => () => {
@@ -69,16 +69,6 @@ export default function BookmarkContainer({ setTabState }) {
 		if (answer) {
 			dispatch(removeAllBookmarkItems());
 		}
-	};
-
-	const handleFilter = ({ target }) => {
-		const inputValue = target.value;
-
-		if (!inputValue) {
-			setSearchTerm("");
-			return;
-		}
-		setSearchTerm(inputValue);
 	};
 
 	const handleEditButton = ({ id, description }) => async () => {
